Stop using custom NServiceBusTransportTest
